Add files via upload
This is an algorithm for the least amount of total litre capacity you need to have in order to travel around towns while refiling it on every destination. In the input file you need to specify both the amount of towns and roads on the first line and then write the connection between towns (x y z) where x and y is the towns that are connected and z the amount of petrol in litre you need to have in order to travel that distance.