Add files via upload
This is a program that calculates the likely outcome in the middle of a voting contest. In the input file you specify on the first line the number of people who have already voted, the number of candidates and the number of people who should vote in total. Then on the second line you need to write the votes that have already happened. The output of the program will show you the number of people who are likely to win the voting contest.